+++ src/main/kotlin/ee/carlrobert/codegpt/ui/textarea/PromptTextField.kt
@@ -102,7 +102,7 @@ class PromptTextField(
 
     private fun getToolWindowHeight(): Int {
         return project.service<ToolWindowManager>()
-            .getToolWindow("CodeGPT")?.component?.visibleRect?.height ?: 400
+            .getToolWindow("ProxyAI")?.component?.visibleRect?.height ?: 400
     }
 }